+++ NEWS.md
@@ -11,6 +11,8 @@ project is using.
 - Added `checks` to run basic checks on packer projects.
 - Added `use_loader_svelte`.
 - Added `add_plugin_pwa` for progressive web applications.
+- Added `add_plugin_jsdoc` for generating documentation from
+JavaScript code à la roxygen2.
 
 ## packer 0.1.0